Use a dedicated `Terminators` class instead of CBs
